+++ .env.development
@@ -2,5 +2,6 @@ NEXT_PUBLIC_MATOMO_URL=https://stats.data.gouv.fr
 NEXT_PUBLIC_MATOMO_SITE_ID=159
 NEXT_PUBLIC_RDV_SOLIDARITES_DEMO_URL=http://localhost:5000
 NEXT_PUBLIC_RDV_SOLIDARITES_PROD_URL=http://localhost:5000
+NEXT_PUBLIC_RDV_SOLIDARITES_USER_PATH=/api/v1/users
 NEXT_PUBLIC_ORGANISATION_ID_DEMO=1
 NEXT_PUBLIC_ORGANISATION_ID_PROD=1
